again

+++ Banca/src/Database.cpp
@@ -12,16 +12,16 @@ void Database::CopyToFile(Banca& banca)
 
 	fout << banca.GetNrClienti() << "\n";
 
-	for (int i = 0; i < clienti.size(); ++i)
+	for (size_t i = 0; i < clienti.size(); ++i)
 	{
 		fout << clienti[i].GetId() << " " << clienti[i].GetName() << " " << clienti[i].GetNrConturi() << " " << clienti[i].GetNrCredite() << " ";
 
 		std::vector<Cont> conturi = clienti[i].GetConturi();
 		std::vector<Credit> credite = clienti[i].GetCredite();
-		for (int j = 0; j < clienti[i].GetNrConturi(); ++j)
+		for (size_t j = 0; j < clienti[i].GetNrConturi(); ++j)
 			fout << conturi[j].GetId() << " " << conturi[j].GetSold() << " ";
 
-		for (int j = 0; j < clienti[i].GetNrCredite(); ++j)
+		for (size_t j = 0; j < clienti[i].GetNrCredite(); ++j)
 			fout << credite[j].GetId() << " " << credite[j].GetSold() << " ";
 	}
